Lets not kill the testplugin

+++ src/MiNET/TestPlugin/NoDamageHealthManager.cs
@@ -9,7 +9,7 @@ namespace TestPlugin
 		{
 		}
 
-		public override void TakeHit(Entity source, int damage = 1, DamageCause cause = DamageCause.Unknown)
+		public override void TakeHit(Entity source, double damage = 1, DamageCause cause = DamageCause.Unknown)
 		{
 			//base.TakeHit(source, 0, cause);
 		}
